Hash raw file and check if its cached file is in the cache directory
At the moment, this is all done in the pitch phase which could be slow for every file. There is some thought at front-loading this work with a plugin or something like that.

Ideally, this function should only be checking a lookup. If it is there, it reads the cached file and processes it. Otherwise, nothing.